service 계층 두개로 분리

+++ src/main/java/com/webtoon/cartoon/domain/Cartoon.java
@@ -18,6 +18,8 @@ import lombok.NoArgsConstructor;
 import javax.persistence.*;
 
 
+import java.util.Objects;
+
 import static com.webtoon.util.constant.Constant.*;
 import static javax.persistence.FetchType.LAZY;
 import static javax.persistence.GenerationType.IDENTITY;
@@ -89,7 +91,7 @@ public class Cartoon extends BaseTimeEntity {
     }
 
     public void validateAuthorityForCartoon(AuthorSession authorSession) {
-        if (author.getId() != authorSession.getId()) {
+        if (author.getId().longValue() != authorSession.getId().longValue()) {
             throw new CartoonForbiddenException();
         }
     }
